requirements.txt
django>=3.2.13,<4.0
psycopg2==2.8.5
boto3==1.14.2
uWSGI==2.0.19.1
requests==2.22.0
django-enumfield==2.0.2
passlib==1.7.2
mozilla-django-oidc==1.2.3
psycopg2-binary==2.8.5
django-bootstrap4~=2.3
whitenoise~=5.2
djangorestframework
django-rest-authtoken
scramp